Add files via upload
ANON is an ALL-IN-ONE fashion store, an e-commerce website designed as part of our internship project. This is purely a frontend project involving HTML, CSS, and JS. Please take a look at it, and if you notice any defects in this project, guide me through your inputs to help me learn.